% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chisq}
\alias{chisq}
\title{Chi Square test}
\usage{
chisq(mor, spfac)
}
\arguments{
\item{mor}{is a vector of MOR data for all species}

\item{spfac}{is the species factor for mor}
}
\value{
a list giving the controlling subgroup (csg) and the tolerance
limit characteristic value (cval)
}
\description{
determine the number of pieces in each species group
with MOR values below the combined group's tolerance limit value and
conduct a Chi Square test at a significance level of 0.01.
}
\details{
See Appendix X7, page 21-22 of ASTM D1990-07.
}
\note{
Follow with iterative Chi Square tests on species subsets if needed.

no continuity correction is applied when computing the test
statistic which results in a warning message
}

