% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{compMethod}
\alias{compMethod}
\alias{compMethod-package}
\title{R package: compMethod}
\description{
Add description here.
}
\note{
Methods tl, chisq, ichisq: Grouping method for Tolerance Limit Properties as outlined in ASTM D1990-07
(file name "D1990 Attachment 1.pdf").  See Section 10.3.1, page 6-7
 General program that handles unequal sample sizes

Methods Tukey and npaov: Grouping method for Mean Properties as outlined in ASTM D1990-07 (file name
"D1990 Attachment 1.pdf").  See Section 10.2.1, page 6
General program that handles unequal sample sizes
}

