% Generated by roxygen2 (4.0.2): do not edit by hand
\name{npaov}
\alias{npaov}
\title{Non-parametric ANOVA}
\usage{
npaov(moe, spfac)
}
\arguments{
\item{moe}{is a vector of MOE data for all species}

\item{spfac}{is the species factor for moe}
}
\value{
a list giving the controlling subgroup (csg) and the mean characteristic value (cval)
}
\description{
Apply a parametric F test of the hypothesis of
equal means between species groups to the ranked MOE data at a
significance level of 0.01.
}
\details{
See Appendix X5, page 20 of ASTM D1990-07.
Follow with a Tukey multiple comparison test if needed.
}
\note{
the mean characteristic value is not weighted by the species
sample sizes
}

