% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tl}
\alias{tl}
\title{tl method}
\usage{
tl(mor, a = 5, b = 75, ...)
}
\arguments{
\item{mor}{is a vector of MOR data for all species}

\item{a}{is the percent of the population below the tolerance limit}

\item{b}{is the percent confidence of the tolerance limit}
}
\value{
tolerance limit
}
\description{
Calculate the order statistic for the a% non-parametric one-sided
tolerance limit with b% confidence
}

