% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Tukey}
\alias{Tukey}
\title{Tukey's multiple comparison test}
\usage{
Tukey(moe, rmoe, spfac, df1, df2, aovtab)
}
\arguments{
\item{moe}{is a vector of sample MOE data for all species}

\item{rmoe}{is the ranked sample MOE data for all species}

\item{spfac}{is the species factor for moe and rmoe}

\item{df1}{is the treatment degrees of freedom for the non-parametric ANOVA test}

\item{df2}{is the error degrees of freedom for the non-parametric ANOVA test}

\item{aovtab}{is the ANOVA table for the non-parametric ANOVA test}
}
\value{
a list giving the controlling subgroup (csg) and the mean
characteristic value (cval)
}
\description{
apply to the ranked data to
determine the group of species that are indistinguishable from the
species with the lowest median (the one with the lowest mean of the
ranked data) at a significance level of 0.01.
}
\details{
See Appendix X6, page 20 of ASTM D1990-07.
}
\note{
the mean characteristic value is not weighted by sample size
Mean differences between each species and the one with the lowest mean
}

